+++ tests/esm.mjs
@@ -4,6 +4,10 @@ import TOML from "@aduh95/toml";
 
 assert.deepStrictEqual(
   TOML.parse(readFileSync(new URL("input.toml", import.meta.url), "utf8")),
+  JSON.parse(readFileSync(new URL("output.json", import.meta.url), "utf8"))
+);
+assert.deepStrictEqual(
+  TOML.parse(readFileSync(new URL("input.toml", import.meta.url))),
   JSON.parse(readFileSync(new URL("output.json", import.meta.url)))
 );